+++ src/main/java/hr/in2/postenipoduzetnikevents/model/OrgUnitType.java
@@ -9,6 +9,11 @@ import javax.persistence.Table;
 
 import lombok.Data;
 
+/**
+ * Entity klasa koja odgovara TIP_ORGANIZACIJSKE_JEDINICE tablici
+ * @author Nikola Gereci
+ * @since 0.0.1-SNAPSHOT
+ */
 @Entity
 @Data
 @Table(name = "TIP_ORGANIZACIJSKE_JEDINICE")
@@ -16,7 +21,7 @@ public class OrgUnitType {
 	@Id
 	@GeneratedValue
 	@Column(name = "ID")
-	private int id;
+	private long id;
 	
 	@Column(name = "NAZIV")
 	private String name;
